# model: introduce separate types for JSON encoding