Introduce an internal package with types for encoding to Elasticsearch document JSON structure, remove mapStr conversion. The conversion between these types is manual, just like the old "fields" methods that are now removed.

There is a structural change to the metricset type. Rather than adding fields directly to the top of the output, we now nest metrics under `metricset.samples`. This will need an accompanying change to the metric data stream ingest pipelines.

The JSON encoding methods are generated by fastjson. I've added a simple benchmark, and compared against main:

```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-data/model
cpu: AMD Ryzen 7 PRO 5850U with Radeon Graphics
                   │ /tmp/old.txt │            /tmp/new.txt            │
                   │    sec/op    │   sec/op     vs base               │
MarshalFastJSON-16   20.396µ ± 4%   6.278µ ± 3%  -69.22% (p=0.002 n=6)

                   │ /tmp/old.txt  │            /tmp/new.txt
│
                   │     B/op      │     B/op      vs base
│
MarshalFastJSON-16   12.332Ki ± 0%   1.844Ki ± 0%  -85.05% (p=0.002 n=6)

                   │ /tmp/old.txt │           /tmp/new.txt            │
                   │  allocs/op   │ allocs/op   vs base               │
MarshalFastJSON-16    116.00 ± 0%   31.00 ± 0%  -73.28% (p=0.002 n=6)
``