@simitt 2023-05-11
>There is a structural change to the metricset type. Rather than adding fields directly to the top of the output, we now nest metrics under metricset.samples. This will need an accompanying change to the metric data stream ingest pipelines.

Do I understand this correct that you are planning on adding an ingest pipeline to the apm package, which ensures that the metrics are set to top level again and not cause any changes in the ES data streams?