@axw 2023-05-11
> Do I understand this correct that you are planning on adding an ingest pipeline to the apm package, which ensures that the metrics are set to top level again and not cause any changes in the ES data streams?

Correct, already done in https://github.com/elastic/apm-server/pull/10792